Fix distance calculation in SquashFS when block boundaries are exceeded
As stated in the documentation `Inodes are packed into the metadata blocks, and are not aligned to block boundaries, therefore inodes overlap compressed blocks` it is possible, that metadata blocks are not within the data block.

The code in general accounts for this situation. However the distance calculation worked with incorrect assumptions in regards to this.

This leads to situations, where for example directory metadata blocks could not be calculated correctly, when the metadata spans across multiple data blocks.